+++ .gitignore
@@ -22,6 +22,4 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 target\
-target\
-target\
 compiled\
